feat: Instalação de dotenv
